+++ Plugin/src/Content/Maps/IndustrialFan.cs
@@ -13,6 +13,11 @@ public class IndustrialFan : NetworkBehaviour
     public float suctionForce = 15f;
     public ParticleSystem redMistEffect = null!;
 
+    private void Start()
+    {
+        suctionForce = 6;
+    }
+
     private void Update()
     {
         if (!IsServer) return;